+++ Pandora/Pattern/Object/Group.hs
@@ -4,10 +4,10 @@ import Pandora.Pattern.Object.Quasiring (Quasiring)
 
 {- |
 > When providing a new instance, you should ensure it satisfies the two laws:
-> * Right absorption: x * inverse x ≡ zero
-> * Left absorption: inverse x * x ≡ zero
+> * Right absorption: x * invert x ≡ zero
+> * Left absorption: invert x * x ≡ zero
 -}
 
 class Quasiring a => Group a where
-	{-# MINIMAL inverse #-}
-	inverse :: a -> a
+	{-# MINIMAL invert #-}
+	invert :: a -> a