+++ CHANGELOG.md
@@ -181,3 +181,4 @@
 * Define precedence for `*` and `+`
 * Change superclass of `Group` class - `Quasiring` instead of `Monoid`
 * Rename `Jet`'s constructor to `Jet` (previously - `:-`)
+* Rename `Group`'s method from `inverse` to `invert`